+++ pyxel/models/charge_generation/photoelectrons.py
@@ -6,11 +6,10 @@
 #  the terms contained in the file ‘LICENCE.txt’.
 
 """Simple model to convert photon into photo-electrons inside detector."""
-import logging
-
 import numpy as np
 
 from pyxel.detectors import Detector
+
 
 # TODO: docstring, private function, what is eta,