+++ pyxel/util/__init__.py
@@ -19,6 +19,7 @@ from pyxel.util.memory import get_size, memory_usage_details
 from pyxel.util.examples import download_examples
 from pyxel.util.timing import time_pipeline
 from pyxel.util.add_model import create_model
+from pyxel.util.random import temporary_random_state
 
 __all__ = [
     "convert_to_int",